Placing common code into a single file.
